// models/IProduct.ts
export interface IProduct {
    about: string;
    backgroundcolor: string;
    categories: string;
    category_type: string;
    color: string;
    description: string;
    id: number;
    image: string;
    image_arr: string[];
    isLogo: number;
    isNew: number;
    isTop: number;
    isVisible: number;
    recommend: string;
    sizeGrid: string;
    sizeImg: string[];
    subColor: string;
    title: string;
    uniqCode: number;
}